# docs/api/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_GriffinGuard_Griffino_internal_store.PluginStatus:
    enum:
    - pending_setup
    - ready
    - pulling
    - starting
    - running
    - stopped
    - failed
    type: string
    x-enum-comments:
      StatusFailed: Start failed / 启动失败
      StatusPendingSetup: Manifest downloaded, config not yet done / 元文件已下载，未完成配置
      StatusPulling: Pulling images / 正在拉取镜像
      StatusReady: Config done, never started / 配置完成，从未启动
      StatusRunning: Running / 运行中
      StatusStarting: Images ready, container starting / 镜像就绪，容器启动中
      StatusStopped: Stopped / 已停止
    x-enum-varnames:
    - StatusPendingSetup
    - StatusReady
    - StatusPulling
    - StatusStarting
    - StatusRunning
    - StatusStopped
    - StatusFailed
...
  internal_api.AppError:
    properties:
      code:
        $ref: '#/definitions/internal_api.ErrorCode'
      detail: {}
      message:
        type: string
    type: object
  internal_api.ErrorCode:
    enum:
    - PLUGIN_NOT_FOUND
    - PLUGIN_ALREADY_INSTALLED
    - PLUGIN_ALREADY_RUNNING
    - PLUGIN_NOT_RUNNING
    - PLUGIN_NOT_CONFIGURED
    - PLUGIN_INVALID_STATE
    - PLUGIN_START_FAILED
    - PLUGIN_STOP_FAILED
    - PLUGIN_UNINSTALL_FAILED
    - PLUGIN_LOAD_FAILED
    - PLUGIN_SAVE_FAILED
    - PLUGIN_IMAGE_NOT_ALLOWED
    - PLUGIN_LIST_FAILED
    - REGISTRY_FETCH_FAILED
    - REGISTRY_DOWNLOAD_FAILED
    - REGISTRY_VERSION_NOT_FOUND
    - PLUGIN_ALREADY_UP_TO_DATE
    - PLUGIN_UPGRADE_FAILED
    - AUTH_INVALID_REQUEST
    - AUTH_INVALID_CREDENTIALS
    - AUTH_TOKEN_INVALID
    - AUTH_NOT_LOGGED_IN
    - AUTH_PERMISSION_DENIED
    - AUTH_RATE_LIMITED
    - AUTH_SESSION_FAILED
    - AUTH_PASSWORD_TOO_SHORT
    - AUTH_WRONG_PASSWORD
    - AUTH_PASSWORD_HASH_FAILED
    - AUTH_PASSWORD_SAVE_FAILED
    - USER_NOT_FOUND
    - USER_ALREADY_EXISTS
    - USER_INVALID_REQUEST
    - USER_CANNOT_SELF_DISABLE
    - USER_CANNOT_SELF_DELETE
    - USER_SAVE_FAILED
    - USER_DELETE_FAILED
    - USER_LIST_FAILED
    - ROUTE_INVALID_REQUEST
    - ROUTE_FETCH_FAILED
    - ROUTE_SAVE_FAILED
    - BLUEPRINT_NOT_FOUND
    - BLUEPRINT_INVALID_REQUEST
    - BLUEPRINT_SAVE_FAILED
    - BLUEPRINT_DELETE_FAILED
    - BLUEPRINT_FETCH_FAILED
    - BLUEPRINT_PORT_MISMATCH
    - ACTION_NOT_FOUND
    - ACTION_INVALID_REQUEST
    - ACTION_RATE_LIMITED
    - ACTION_SEND_FAILED
    - USER_CONFIG_INVALID_REQUEST
    - USER_CONFIG_FETCH_FAILED
    - USER_CONFIG_SAVE_FAILED
    - SYSTEM_NOT_INITIALIZED
    - SYSTEM_INTERNAL_ERROR
    - PLUGIN_LOG_NOT_FOUND
    - WIDGET_NOT_FOUND
    - WIDGET_DATA_FAILED
    - DASHBOARD_INVALID_REQUEST
    - DASHBOARD_SAVE_FAILED
    - TASK_NOT_FOUND
    - TASK_FETCH_FAILED
    - SETTINGS_INVALID_REQUEST
    - SETTINGS_SAVE_FAILED
    - SETTINGS_FETCH_FAILED
    - SESSION_NOT_FOUND
    - SESSION_LIST_FAILED
    - USERNAME_TAKEN
    - USERNAME_INVALID
    - USER_ALREADY_DISABLED
    - USER_ALREADY_ENABLED
    - CANNOT_DISABLE_SELF
    - CANNOT_DELETE_LAST_ADMIN
    - CANNOT_DEMOTE_LAST_ADMIN
    - ACCOUNT_DISABLED
    - PASSWORD_REUSE_FORBIDDEN
    - EMAIL_INVALID
    - EMAIL_ALREADY_VERIFIED
    - SETUP_ALREADY_COMPLETED
    - SETUP_NOT_COMPLETED
    - LICENSE_NOT_ACCEPTED
    - ADMIN_ALREADY_EXISTS
    - SMTP_CONNECTION_FAILED
    - SMTP_AUTH_FAILED
    - SMTP_SEND_FAILED
    - CONFIG_KEY_INVALID
    - CONFIG_VALUE_INVALID
    type: string
    x-enum-varnames:
    - ErrPluginNotFound
    - ErrPluginAlreadyInstalled
    - ErrPluginAlreadyRunning
    - ErrPluginNotRunning
    - ErrPluginNotConfigured
    - ErrPluginInvalidState
    - ErrPluginStartFailed
    - ErrPluginStopFailed
    - ErrPluginUninstallFailed
    - ErrPluginLoadFailed
    - ErrPluginSaveFailed
    - ErrPluginImageNotAllowed
    - ErrPluginListFailed
    - ErrRegistryFetchFailed
    - ErrRegistryDownloadFailed
    - ErrRegistryVersionNotFound
    - ErrPluginAlreadyUpToDate
    - ErrPluginUpgradeFailed
    - ErrAuthInvalidRequest
    - ErrAuthInvalidCredentials
    - ErrAuthTokenInvalid
    - ErrAuthNotLoggedIn
    - ErrAuthPermissionDenied
    - ErrAuthRateLimited
    - ErrAuthSessionFailed
    - ErrAuthPasswordTooShort
    - ErrAuthWrongPassword
    - ErrAuthPasswordHashFailed
    - ErrAuthPasswordSaveFailed
    - ErrUserNotFound
    - ErrUserAlreadyExists
    - ErrUserInvalidRequest
    - ErrUserCannotSelfDisable
    - ErrUserCannotSelfDelete
    - ErrUserSaveFailed
    - ErrUserDeleteFailed
    - ErrUserListFailed
    - ErrRouteInvalidRequest
    - ErrRouteFetchFailed
    - ErrRouteSaveFailed
    - ErrBlueprintNotFound
    - ErrBlueprintInvalidRequest
    - ErrBlueprintSaveFailed
    - ErrBlueprintDeleteFailed
    - ErrBlueprintFetchFailed
    - ErrBlueprintPortMismatch
    - ErrActionNotFound
    - ErrActionInvalidRequest
    - ErrActionRateLimited
    - ErrActionSendFailed
    - ErrUserConfigInvalidRequest
    - ErrUserConfigFetchFailed
    - ErrUserConfigSaveFailed
    - ErrSystemNotInitialized
    - ErrSystemInternal
    - ErrPluginLogNotFound
    - ErrWidgetNotFound
    - ErrWidgetDataFailed
    - ErrDashboardInvalidRequest
    - ErrDashboardSaveFailed
    - ErrTaskNotFound
    - ErrTaskFetchFailed
    - ErrSettingsInvalidRequest
    - ErrSettingsSaveFailed
    - ErrSettingsFetchFailed
    - ErrSessionNotFound
    - ErrSessionListFailed
    - ErrUsernameTaken
    - ErrUsernameInvalid
    - ErrUserAlreadyDisabled
    - ErrUserAlreadyEnabled
    - ErrCannotDisableSelf
    - ErrCannotDeleteLastAdmin
    - ErrCannotDemoteLastAdmin
    - ErrAccountDisabled
    - ErrPasswordReuseForbidden
    - ErrEmailInvalid
    - ErrEmailAlreadyVerified
    - ErrSetupAlreadyCompleted
    - ErrSetupNotCompleted
    - ErrLicenseNotAccepted
    - ErrAdminAlreadyExists
    - ErrSMTPConnectionFailed
    - ErrSMTPAuthFailed
    - ErrSMTPSendFailed
    - ErrConfigKeyInvalid
    - ErrConfigValueInvalid
  internal_api.PluginDTO:
    properties:
      configDirty:
        type: boolean
      failReason:
        type: string
      failStage:
        type: string
      id:
        type: string
      installedAt:
        type: string
      isDevPlugin:
        type: boolean
      pluginDir:
        type: string
      runtimeInfo:
        $ref: '#/definitions/internal_api.RuntimeInfoDTO'
      status:
        $ref: '#/definitions/github_com_GriffinGuard_Griffino_internal_store.PluginStatus'
    type: object
  internal_api.RuntimeInfoDTO:
    properties:
      containers:
        additionalProperties:
          type: string
        type: object
      network:
        type: string
      rabbitmqUser:
        type: string
      redisUser:
        type: string
    type: object
host: 127.0.0.1:7070
info:
  contact: {}
  description: Local-only control-plane API for the Griffino daemon. The API binds
    to 127.0.0.1:7070 and is consumed by the embedded web console. All non-public
    endpoints require a bearer session token obtained from POST /auth/login.
  title: Griffino API
  version: "1.0"
paths:
...
  /plugins:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/internal_api.PluginDTO'
              type: array
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.AppError'
      security:
      - BearerAuth: []
      summary: List installed plugins
      tags:
      - Plugins
  /plugins/{id}:
    delete:
      parameters:
      - description: Plugin ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.AppError'
      security:
      - BearerAuth: []
      summary: Uninstall a plugin
      tags:
      - Plugins
    get:
      parameters:
      - description: Plugin ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.PluginDTO'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api.AppError'
      security:
      - BearerAuth: []
      summary: Get a plugin
      tags:
      - Plugins
...
  /plugins/{id}/config:
    get:
      parameters:
      - description: Plugin ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.AppError'
      security:
      - BearerAuth: []
      summary: Get plugin admin config schema and values
      tags:
      - Plugins
    post:
      consumes:
      - application/json
      parameters:
      - description: Plugin ID
        in: path
        name: id
        required: true
        type: string
      - description: config map and action (save | save_and_start | save_and_restart)
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "202":
          description: Accepted
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api.AppError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_api.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.AppError'
      security:
      - BearerAuth: []
      summary: Save plugin admin config
      tags:
      - Plugins
  /plugins/{id}/start:
    post:
      parameters:
      - description: Plugin ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.AppError'
      security:
      - BearerAuth: []
      summary: Start a plugin
      tags:
      - Plugins
  /plugins/{id}/stop:
    post:
      parameters:
      - description: Plugin ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.AppError'
      security:
      - BearerAuth: []
      summary: Stop a plugin
      tags:
      - Plugins
...
securityDefinitions:
  BearerAuth:
    description: Session token issued by POST /auth/login, sent as "Bearer <token>".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"